+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## **1.22.0** (2021. 03. 13.)
+## **1.22.0** (2021. 03. 13.)  [_Sort All The Things Edition_]
 - Add commands to convert between decimal and hexadecimal numbers using 8, 16, 32 and 64 bit arithmetic.
 - Ignore empty and whitespace only lines when converting between decimal and hexadecimal numbers.
 - Add _Sort lines ascending/descending (case sensitive)_ commands
 - Add _Sort lines ascending/descending (case sensitive/insensitive at column of selection start)_ commands
+- Add _Sort lines by length (and then case sensitive/insensitive)_ commands
 - Add _Shuffle lines_ command
 
 ## **1.21.0** (2021. 02. 14.)